refactor: move skip to internal
