+++ internal/services/azapi_resource.go
@@ -27,7 +27,7 @@ import (
 	"github.com/Azure/terraform-provider-azapi/internal/services/myvalidator"
 	"github.com/Azure/terraform-provider-azapi/internal/services/parse"
 	"github.com/Azure/terraform-provider-azapi/internal/services/preflight"
-	"github.com/Azure/terraform-provider-azapi/internal/services/skip"
+	"github.com/Azure/terraform-provider-azapi/internal/skip"
 	"github.com/Azure/terraform-provider-azapi/internal/tf"
 	"github.com/Azure/terraform-provider-azapi/utils"
 	"github.com/cenkalti/backoff/v4"
@@ -591,8 +591,12 @@ func (r *AzapiResource) Create(ctx context.Context, request resource.CreateReque
 func (r *AzapiResource) Update(ctx context.Context, request resource.UpdateRequest, response *resource.UpdateResponse) {
 	// See if we can skip the external API call (changes are to state only)
 	var plan, state AzapiResourceModel
-	request.State.Get(ctx, &state)
-	request.Plan.Get(ctx, &plan)
+	if response.Diagnostics.Append(request.Plan.Get(ctx, &plan)...); response.Diagnostics.HasError() {
+		return
+	}
+	if response.Diagnostics.Append(request.State.Get(ctx, &state)...); response.Diagnostics.HasError() {
+		return
+	}
 	if skip.CanSkipExternalRequest(plan, state, "update") {
 		response.Diagnostics.Append(response.State.Set(ctx, plan)...)
 		tflog.Debug(ctx, "azapi_resource.CreateUpdate skipping external request as no unskippable changes were detected")
